Updates to tests and moving items around
